Minor changes in PHPUnit Tests.
