+++ tests/api/AuthLoginTest.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace Tests\api;
+
+class AuthLoginTest extends BaseTestCase
+{
+    /**
+     * Test user login endpoint and get a JWT Bearer Authorization.
+     */
+    public function testLogin()
+    {
+        $response = $this->runApp('POST', '/login', ['email' => 'test@user.com', 'password' => 'AnyPass1000']);
+
+        $result = (string) $response->getBody();
+
+        self::$jwt = json_decode($result)->message->Authorization;
+
+        $this->assertEquals(200, $response->getStatusCode());
+        $this->assertStringContainsString('Authorization', $result);
+        $this->assertStringContainsString('Bearer', $result);
+    }
+}
